services/ocr-service/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
alembic>=1.12.0
redis>=5.0.0
celery>=5.3.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytesseract>=0.3.10
pillow>=10.0.0
pydantic>=2.5.0
python-multipart>=0.0.6
prometheus-client>=0.19.0
pyyaml>=6.0.1
httpx>=0.27.0
psycopg2-binary>=2.9.9  # Added for PostgreSQL connection testing